pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fuzzy_actions"
version = "0.1.0"
description = "DQN training on gridworld_env tasks"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "gymnasium>=0.28.0",
    "tensorboard>=2.14.0",
    "pillow>=9.0.0",
    # gridworld-env must be installed separately from its source directory:
    #   pip install -e ../gridworld_env
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["fuzzy_actions"]
